functions/requirements.txt
firebase-functions
firebase-admin
google-cloud-firestore
openai
pinecone
python-dotenv
moviepy
pillow
numpy
requests
